+++ data/books.json
@@ -1,1 +1,1 @@
-[{"title":"Gifted Hands","publisher":"Ben Carson","cover_state":"good","publish_date":"1922-01-01","label":{"title":"Health","color":"Green"}},{"title":"Casflow Quadrant","publisher":"Robert T. Kiyosaki","cover_state":"good","publish_date":"1822-01-01","label":{"title":"Wealth","color":"Blue"}},{"title":"Rich Dad Poor Dad","publisher":"Robert T. Kiyosaki","cover_state":"good","publish_date":"2823-01-01","label":{"title":"Finance","color":"Blue"}},{"title":"The Rules of Work","publisher":"Richard Templar","cover_state":"good","publish_date":"2828-01-05","label":{"title":"Ethics","color":"Brown"}},{"title":"the end","publisher":"anuar","cover_state":"good","publish_date":"2022-02-02","label":{"title":"End","color":"blue"}}]
+[{"title":"Gifted Hands","publisher":"Ben Carson","cover_state":"good","publish_date":"1922-01-01","label":{"title":"Health","color":"Green"}},{"title":"Casflow Quadrant","publisher":"Robert T. Kiyosaki","cover_state":"good","publish_date":"1822-01-01","label":{"title":"Wealth","color":"Blue"}},{"title":"Rich Dad Poor Dad","publisher":"Robert T. Kiyosaki","cover_state":"good","publish_date":"2823-01-01","label":{"title":"Finance","color":"Blue"}},{"title":"The Rules of Work","publisher":"Richard Templar","cover_state":"good","publish_date":"2828-01-05","label":{"title":"Ethics","color":"Brown"}},{"title":"the end","publisher":"anuar","cover_state":"good","publish_date":"2022-02-02","label":{"title":"End","color":"blue"}},{"title":"Harry Potter","publisher":"idk","cover_state":"bad","publish_date":"2000-01-01","label":{"title":"idk","color":"gold"}}]
