+++ data/games.json
@@ -1,1 +1,1 @@
-[{"multiplayer":"multiplayer","publish_date":"2015-02-02","last_played_at":"2022-02-02","author":{"first_name":"anuar","last_name":"sha"}},{"multiplayer":"ricky","publish_date":"2020-02-02","last_played_at":"2022-02-02","author":{"first_name":"ricky","last_name":"mormor"}}]
+[{"multiplayer":"multiplayer","publish_date":"2015-02-02","last_played_at":"2022-02-02","author":{"first_name":"anuar","last_name":"sha"}},{"multiplayer":"ricky","publish_date":"2020-02-02","last_played_at":"2022-02-02","author":{"first_name":"ricky","last_name":"mormor"}},{"multiplayer":"idk","publish_date":"1998-12-31","last_played_at":"2022-05-09","author":{"first_name":"kanye","last_name":"west"}}]
