+++ data/genres.json
@@ -1,18 +1,22 @@
 [
   {
-    "GenreId": 832,
+    "GenreId": 871,
     "GenreName": "fremulon"
   },
   {
-    "GenreId": 56,
+    "GenreId": 892,
     "GenreName": "then"
   },
   {
-    "GenreId": 75,
+    "GenreId": 705,
     "GenreName": "thewtre"
   },
   {
-    "GenreId": 90,
+    "GenreId": 88,
     "GenreName": "rock"
+  },
+  {
+    "GenreId": 402,
+    "GenreName": "RnB"
   }
 ]
