+++ data/music_albums.json
@@ -14,5 +14,9 @@
   {
     "on_spotify": true,
     "Publish_date": "2022-02-02"
+  },
+  {
+    "on_spotify": false,
+    "Publish_date": "2019-03-03"
   }
 ]
